code-generator/requirements.txt
black==25.9.0
click==8.3.0
isort==7.0.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
python-dotenv==1.2.1
pytokens==0.2.0
